requirements.txt
streamlit
sys
google.generativeai
streamlit_folium
folium
requests
geopy.geocoders